move zero values fixer into helpers module; add broken serialization test
