+++ package.json
@@ -61,7 +61,7 @@
   "scripts": {
     "test": "grunt test",
     "proto": "pbjs --keep-case -t static-module ./proto/protocol.proto -o ./proto/protocol.js",
-    "proto:test": "pbjs --keep-case -t static-module ./proto/protocol.proto ./test/sources/proto/cryptocurrency.proto -o ./test/sources/proto/cryptocurrency.js",
+    "proto:test": "pbjs --keep-case -t static-module ./proto/protocol.proto ./test/sources/proto/**/*.proto -o ./test/sources/proto/stubs.js",
     "integration:build": "cargo build --manifest-path integration-tests/Cargo.toml",
     "integration": "mocha -r babel-register integration-tests/test.js",
     "preintegration:unix": "npm run integration:build && npm run postintegration:unix && cargo run --manifest-path integration-tests/Cargo.toml & sleep 10",
