+++ src/types/generic.js
@@ -1,6 +1,6 @@
 import * as crypto from '../crypto'
+import { cleanZeroValuedFields } from '../helpers'
 
-const intTypes = ['sint32', 'uint32', 'int32', 'sfixed32', 'fixed32', 'sint64', 'uint64', 'int64', 'sfixed64', 'fixed64']
 /**
  * @constructor
  * @param {Object} schema
@@ -11,34 +11,12 @@ class Type {
   }
 
   /**
-   * @param {Type} schema
-   * @param {Object} data
-   * @param {Object} object
-   * @returns {Object}
-   */
-  fixZeroIntFields (schema, data, object) {
-    const keys = Object.keys(data)
-    keys.forEach(element => {
-      if (schema.fields && schema.fields[element] && schema.fields[element].name && schema.fields[element].type) {
-        if (schema.fields[element].type === 'message') {
-          object[element] = this.fixZeroIntFields(schema.fields[element], data[element], object)
-        }
-        if (!(intTypes.find((value) => { return value === schema.fields[element].type }) && data[element] === 0)) {
-          object[element] = data[element]
-        }
-      }
-    })
-
-    return data
-  }
-
-  /**
    * Serialize data into array of 8-bit integers
    * @param {Object} data
    * @returns {Array}
    */
   serialize (data) {
-    const object = this.fixZeroIntFields(this.schema, data, {})
+    const object = cleanZeroValuedFields(this.schema, data, {})
     return Array.from(this.schema.encode(object).finish())
   }
