+++ test/sources/data/map-proof.json
@@ -159,5 +159,48 @@
         }
       ]
     }
+  },
+
+  "valid-timestamp": {
+    "expected": {
+      "valueType": "TimestampEntry",
+      "merkleRoot": "e8d33d340c734f2d8902fd029eaae8ab9ad23260d2507fe377a2baef562f4416",
+      "entries": [
+        [
+          "17fbc4ce1b20713fc29bbc889bb6bad24978bc243d7983db3c53ef6bfa54cb24",
+          {
+            "time": {
+              "nanos": 964216000,
+              "seconds": 1544628830
+            },
+            "timestamp": {
+              "content_hash": { "data": [23, 251, 196, 206, 27, 32, 113, 63, 194, 155, 188, 136, 155, 182, 186, 210, 73, 120, 188, 36, 61, 121, 131, 219, 60, 83, 239, 107, 250, 84, 203, 36] },
+              "metadata": ""
+            },
+            "tx_hash": { "data": [117, 145, 119, 201, 77, 246, 230, 2, 107, 25, 12, 42, 138, 78, 82, 101, 68, 214, 30, 54, 99, 229, 186, 201, 4, 82, 19, 205, 12, 158, 226, 60] }
+          }
+        ]
+      ]
+    },
+
+    "data": {
+      "proof": [],
+      "entries": [
+        {
+          "key": "17fbc4ce1b20713fc29bbc889bb6bad24978bc243d7983db3c53ef6bfa54cb24",
+          "value": {
+            "time": {
+              "nanos": 964216000,
+              "seconds": 1544628830
+            },
+            "timestamp": {
+              "content_hash": { "data": [23, 251, 196, 206, 27, 32, 113, 63, 194, 155, 188, 136, 155, 182, 186, 210, 73, 120, 188, 36, 61, 121, 131, 219, 60, 83, 239, 107, 250, 84, 203, 36] },
+              "metadata": ""
+            },
+            "tx_hash": { "data": [117, 145, 119, 201, 77, 246, 230, 2, 107, 25, 12, 42, 138, 78, 82, 101, 68, 214, 30, 54, 99, 229, 186, 201, 4, 82, 19, 205, 12, 158, 226, 60] }
+          }
+        }
+      ]
+    }
   }
 }
